+++ test/memhouse/f1_ash_domain_backbone_test.exs
@@ -77,6 +77,7 @@ defmodule MemHouse.F1AshDomainBackboneTest do
     MemHouse.Operations.DreamTimeWatermark,
     MemHouse.Operations.PipelineRun,
     MemHouse.Operations.UsageEvent,
+    MemHouse.Retrieval.RecallDocument,
     MemHouse.Retrieval.RetrievalProfile,
     MemHouse.Skills.SkillRequirementCard,
     MemHouse.Topology.RoleGrant,
@@ -94,7 +95,7 @@ defmodule MemHouse.F1AshDomainBackboneTest do
     documents dream_time_watermarks entities entity_mentions erasure_requests external_identities gate_decisions
     governance_gate_rules knowledge_consents knowledge_items knowledge_lifecycle_events
     knowledge_relations messages model_role_configs peer_ask_preferences peer_queries
-    peer_query_deliveries peers pipeline_runs policy_configs projections provenances
+    peer_query_deliveries peers pipeline_runs policy_configs projections provenances recall_documents
     retrieval_profiles role_grants scope_relations scopes session_participants session_scopes
     sessions skill_requirement_cards usage_events validation_items
   )